new interesting concepts

+++ src/current/map/prebuilt/int32/hash.c
@@ -4,5 +4,5 @@ uint64 __map_prebuilt_int32_hash (
 /**
  * This function returns an identity hash value. */
 {
-  return (uint32) *((int32*) key);
+  return (uint64) *((int32*) key);
 }
